Read SPIM_EXCEPTION_HANDLER environment variable on Unix.
git-svn-id: http://svn.code.sf.net/p/spimsimulator/code@322 663c0655-963b-4438-a250-40b5657b8e18

+++ xspim/buttons.c
@@ -583,7 +583,7 @@ static void
 clear_program_state_action (Widget w, XtPointer client_data,
 			    XtPointer call_data)
 {
-  int clear_op = (int) client_data;
+  long clear_op = (long) client_data;
 
   switch (clear_op)
     {
